backend/requirements.txt
blinker==1.6.2
click==8.1.7
cryptography==41.0.4
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
pytest==7.4.2
PyMySQL==1.1.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
flask-smorest==0.42.1
coverage==7.3.2
Flask-WTF==1.2.1
marshmallow==3.20.1
marshmallow-enum==1.5.1
Flask-Principal==0.4.0
python-dateutil==2.8.2
scikit-learn==1.3.2
xlrd==2.0.1
pandas==2.1.3
openpyxl==3.1.2
spacy==3.7.2
